Add new route path to display and create new message.
In this commit, I added a new route path in the routes/new.js file to handle the /new route. The route displays a form where users can enter a new message and submit it. When the form is submitted, the new message is added to the existing messages array in routes/index.js and displayed on the index page. Additionally, I added a button on the index page that links to the new route, providing an easy way to access the form for adding new messages.